+++ CHANGELOG.md
@@ -32,6 +32,7 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
   [#139](https://github.com/hynek/svcs/pull/139)
 
 - Support for Python 3.9.
+  [#152](https://github.com/hynek/svcs/pull/152)
 
 
 ### Changed