+++ CHANGELOG.md
@@ -31,6 +31,9 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
   [#135](https://github.com/hynek/svcs/discussions/135)
   [#139](https://github.com/hynek/svcs/pull/139)
 
+- Support for Python 3.9.
+  [#152](https://github.com/hynek/svcs/pull/152)
+
 
 ### Changed